Move UploadIntent tests to uploader_subcommand_test.py
