+++ tensorboard/uploader/BUILD
@@ -123,7 +123,6 @@ py_test(
         ":test_util",
         ":upload_tracker",
         ":uploader",
-        ":uploader_subcommand",
         ":util",
         "//tensorboard:data_compat",
         "//tensorboard:dataclass_compat",
@@ -142,6 +141,18 @@ py_test(
         "//tensorboard/util:test_util",
         "@com_google_protobuf//:protobuf_python",
         "@org_pythonhosted_mock",
+    ],
+)
+
+py_test(
+    name = "uploader_subcommand_test",
+    srcs = ["uploader_subcommand_test.py"],
+    srcs_version = "PY3",
+    deps = [
+        ":dry_run_stubs",
+        ":server_info",
+        ":uploader",
+        ":uploader_subcommand",
     ],
 )
