+++ tensorboard/uploader/uploader_test.py
@@ -43,13 +43,10 @@ from tensorboard.uploader.proto import scalar_pb2
 from tensorboard.uploader.proto import server_info_pb2
 from tensorboard.uploader.proto import write_service_pb2
 from tensorboard.uploader.proto import write_service_pb2_grpc
-from tensorboard.uploader import dry_run_stubs
 from tensorboard.uploader import test_util
 from tensorboard.uploader import upload_tracker
 from tensorboard.uploader import uploader as uploader_lib
-from tensorboard.uploader import uploader_subcommand
 from tensorboard.uploader import logdir_loader
-from tensorboard.uploader import server_info as server_info_lib
 from tensorboard.uploader import util
 from tensorboard.compat.proto import event_pb2
 from tensorboard.compat.proto import graph_pb2
@@ -1987,87 +1984,6 @@ class VarintCostTest(tf.test.TestCase):
         self.assertEqual(uploader_lib._varint_cost(128), 2)
         self.assertEqual(uploader_lib._varint_cost(128 * 128 - 1), 2)
         self.assertEqual(uploader_lib._varint_cost(128 * 128), 3)
-
-
-class UploadIntentTest(tf.test.TestCase):
-    def testUploadIntentUnderDryRunOneShot(self):
-        """Test the upload intent under the dry-run + one-shot mode."""
-        mock_server_info = mock.MagicMock()
-        mock_channel = mock.MagicMock()
-        upload_limits = server_info_pb2.UploadLimits(
-            max_scalar_request_size=128000,
-            max_tensor_request_size=128000,
-            max_tensor_point_size=11111,
-            max_blob_request_size=128000,
-            max_blob_size=128000,
-        )
-        mock_stdout_write = mock.MagicMock()
-        with mock.patch.object(
-            server_info_lib,
-            "allowed_plugins",
-            return_value=_SCALARS_HISTOGRAMS_AND_GRAPHS,
-        ), mock.patch.object(
-            server_info_lib, "upload_limits", return_value=upload_limits
-        ), mock.patch.object(
-            sys.stdout, "write", mock_stdout_write
-        ), mock.patch.object(
-            dry_run_stubs,
-            "DryRunTensorBoardWriterStub",
-            side_effect=dry_run_stubs.DryRunTensorBoardWriterStub,
-        ) as mock_dry_run_stub:
-            intent = uploader_subcommand.UploadIntent(
-                self.get_temp_dir(), dry_run=True, one_shot=True
-            )
-            intent.execute(mock_server_info, mock_channel)
-        self.assertEqual(mock_dry_run_stub.call_count, 1)
-        self.assertRegex(
-            mock_stdout_write.call_args_list[-2][0][0],
-            ".*Done scanning logdir.*",
-        )
-        self.assertEqual(
-            mock_stdout_write.call_args_list[-1][0][0], "\nDone.\n"
-        )
-
-    def testUploadIntentDryRunNonOneShotInterrupted(self):
-        mock_server_info = mock.MagicMock()
-        mock_channel = mock.MagicMock()
-        mock_stdout_write = mock.MagicMock()
-        mock_uploader = mock.MagicMock()
-        with mock.patch.object(
-            mock_uploader, "start_uploading", side_effect=KeyboardInterrupt(),
-        ), mock.patch.object(
-            uploader_lib, "TensorBoardUploader", return_value=mock_uploader
-        ), mock.patch.object(
-            sys.stdout, "write", mock_stdout_write
-        ):
-            intent = uploader_subcommand.UploadIntent(
-                self.get_temp_dir(), dry_run=True, one_shot=False
-            )
-            intent.execute(mock_server_info, mock_channel)
-        self.assertEqual(
-            mock_stdout_write.call_args_list[-1][0][0], "\nInterrupted.\n"
-        )
-
-    def testUploadIntentNonDryRunNonOneShotInterrupted(self):
-        mock_server_info = mock.MagicMock()
-        mock_channel = mock.MagicMock()
-        mock_stdout_write = mock.MagicMock()
-        mock_uploader = mock.MagicMock()
-        with mock.patch.object(
-            mock_uploader, "start_uploading", side_effect=KeyboardInterrupt(),
-        ), mock.patch.object(
-            uploader_lib, "TensorBoardUploader", return_value=mock_uploader
-        ), mock.patch.object(
-            sys.stdout, "write", mock_stdout_write
-        ):
-            intent = uploader_subcommand.UploadIntent(
-                self.get_temp_dir(), dry_run=False, one_shot=False
-            )
-            intent.execute(mock_server_info, mock_channel)
-        self.assertIn(
-            "\nInterrupted. View your TensorBoard at ",
-            mock_stdout_write.call_args_list[-1][0][0],
-        )
 
 
 def _clear_wall_times(request):
